Implemented new validation logic Removed unneeded text in controller Removed all null checks from controllers Employee age search givenInteger -> givenAge Added custom NotFoundException Changed employee age int -> Long Updated ageValidator to reflect employee change Completely redone ExceptionHandlingController Removed custom ExceptionResponse Update employee dto age int -> long Changed all int values in Task to Long Removed fetchType from task Updated task dto int -> Long Updated EmployeeRepository because of age change Updated EmployeeService int -> Long Removed most null checks in serviceImpl Changed NPM -> NotFoundException Updated tests to reflect int -> long changes

+++ src/main/java/ua/slavik/carwash/model/dto/employee/UpdateEmployeeDTO.java
@@ -17,7 +17,7 @@ import javax.validation.constraints.Email;
 public class UpdateEmployeeDTO {
     private Long id;
     @Age
-    private int age;
+    private Long age;
     @StringField(message = "Invalid first name.")
     private String firstName;
     @StringField(message = "Invalid last name.")
