+++ src/main/java/ua/slavik/carwash/model/dto/task/CreateTaskDTO.java
@@ -15,11 +15,11 @@ import javax.validation.constraints.PositiveOrZero;
 @AllArgsConstructor
 public class CreateTaskDTO {
     @PositiveOrZero(message = "Invalid price.")
-    private int price;
+    private Long price;
     @PositiveOrZero(message = "Invalid priority.")
-    private int priority;
+    private Long priority;
     @PositiveOrZero(message = "Invalid duration.")
-    private int duration;
+    private Long duration;
     @StringField(message = "Invalid description.")
     private String description;
     @StringField(message = "Invalid Name.")
